Player refactor
